refactor(dts): split oversized js emit helpers
