+++ crates/tsz-emitter/src/declaration_emitter/core/mod.rs
@@ -9,6 +9,9 @@ mod inferred_predicate_returns;
 mod js_class_static;
 mod js_commonjs_constructor;
 mod js_emit;
+mod js_emit_binding_types;
+mod js_emit_namespace_exports;
+mod js_emit_synthetic;
 mod js_return_type_fallback;
 mod js_synthetic_members;
 mod preamble;
